pyproject.toml
[project]
name = "agent-aided-coding-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "fastmcp>=2.13.1",
    "pytest>=9.0.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
